Encode dist type, bits and rate as ints; clamp distortion params via macros
dist_type carried its enum as a float on the wire and through the delta
queue. Move it, dist_bits and dist_rate to EVENT_TO_DELTA_I and integer
event/config fields - the crusher consumes whole bits and whole sample
counts, so float precision bought nothing there. dist_drive and dist_mix
stay float: they are continuous, and they are the fields a later change
would grow into control-coef vectors.

play_delta's per-param clamp blocks collapse into
DELTA_TO_SYNTH_F_CLAMPED / DELTA_TO_SYNTH_I_CLAMPED; DIST_TYPE stays
special-cased because a type change also resets the crusher state.

Three knock-ons the type change forces:

- The GH parser gates each value on AMY_IS_SET before assigning: the
  float parse used NaN as its absent marker, and NaN-to-int is UB. Bits
  saturates to 24 at the cast for the same reason.
- has_synth_params tested the five dist fields through the all-float
  _RET_TRUE_IF_5_F_SET helper; an int sentinel cast to float is an
  ordinary value, not NaN, so the test went through the per-field
  _Generic form instead.
- The delta-to-event map read the three params back through _CASE_F,
  i.e. the float half of the data union; they move to _CASE_I.

Verification: parse/print round-trip over the G commands (13 cases,
including GH int formatting, off forms, filter+dist in both orders,
out-of-range saturation); SILENT-head clip RMS pin unchanged
(1174 -> 1529 at GC1GD8); enable-then-disable renders byte-identical
to dry. c-api/godot generators re-run with zero drift.

+++ src/amy.h
@@ -610,10 +610,10 @@ typedef struct amy_event {
     uint8_t algorithm;
     uint8_t filter_type;
     // Per-osc distortion ('G' distortion sub-commands on the wire).
-    float dist_type;
+    uint8_t dist_type;
     float dist_drive;
-    float dist_bits;
-    float dist_rate;
+    uint8_t dist_bits;
+    uint16_t dist_rate;
     float dist_mix;
     float eq_l;  // not in synth
     float eq_m;  // not in synth
@@ -666,8 +666,8 @@ typedef struct amy_event {
 typedef struct dist_config {
     uint8_t type;    // One of the DIST_ values.
     float drive;     // Pre-gain, 0..16 (fold depth for DIST_FOLD).
-    float bits;      // DIST_CRUSH bit depth; >= 24 disables quantization.
-    float rate;      // DIST_CRUSH sample-hold length in samples; 1 disables.
+    uint8_t bits;    // DIST_CRUSH bit depth; >= 24 disables quantization.
+    uint16_t rate;   // DIST_CRUSH sample-hold length in samples; 1 disables.
     float mix;       // Wet/dry, 0..1.
 } dist_config_t;
